Like gate
The page that hosts the content, using a div and solely CSS to display the message or image we want to present so the user likes our page before continuing to the application. Can be skipped by editing the index.php file and can be edited by editing the style.css file.